Reorganise job component's job column structure
This essentially tags each column with an underlying variable, which will
be specific to the job.

For instance, such a variable might be some sort of 'grouping' factor, if
there is more than one observation per row (the data are long); such a
variable might relate to time or the sampling frame; or such a variable
might relate to density or infection pressure.

This aims to work around the apparently intractable behaviour of LinkML's
RDF conversion, which does not support ordered RDF collections or
containers (by default or otherwise).

My original idea for this was to create classes which inherit from the
main `JobDesc' class, e.g. `JobMAE' for mean absolute error, using the
`is_a' directive. While I could get these to inherit just fine (i.e. the
`atomic_name' and `title' slots), when loading the RDF manifest back into
LinkML, these would be coerced into the main `JobDesc' class, which
always failed, since it obviously doesn't inherit the other way.

I am not certain about what the mechanism behind this is or if there is a
way to make LinkML accept these sub-classes. There is certainly no way to
specify more than one class as an acceptable range using the `any_of'
directive: https://linkml.io/linkml/schemas/advanced.html#unions-as-ranges

E.g. this is invalid:

     any_of:
	- range: JobMAE
	- range: JobDummy

This has the potential to get annoying for many columns, but at the
moment our model evaluation scripts take very few input columns, so it is
OK to be explicit here.