Separate data-set descriptive attributes from column descriptive
The reasoning behind this is that these slots are more general than
columns, i.e. they are useful both in schemata for data files and in the
data model job component. Equally, there were slots in the 'job'
component which are more general than descriptions of jobs and tables.

Thus, I have moved these slots into the data model's 'core' component.

I have further generalised `job_type' to `type' (analogous to re-using
`description' and other slots), and I also removed the `job_' prefix
from the `scope' slots, since this is implicit given these slots only
appear in the job description class (`JobDesc').

Generalising these slots is productive because these slots were just
duplicating slots from the LinkML meta-model, which usually are just
linked to Dublin Core terms. Just like LinkML's duplicated slots, I set
the slot URI to that of the corresponding Dublin Core terms property.

Furthermore, I have set the range of the `variable' and `table' slots
to `uriorcurie'. This is because these always concern a slot and/or
table declared in the schema or in the data model. For instance, the
table variable is declared in the job manifest, whereas the underlying
variable is declared in either the 'descriptive', 'collected', or
'modelled' component of the data model. (Separating these with a
different prefix for the generated manifest may be productive.)