Add additional prefixes to be generated
The `rap' prefix is a nominal location which serves the job manifest,
job results, job data and possibly certain descriptive statistics,
something which will be specific to the job and shape of the data.

RAP is a civil service term referring generally to statistical and
data pipelines, something which emphasises reproducibility and use of
non-proprietary software. Thus putting outputs of jobs at a location
called this made sense.