Add `prep' option to webroot copying scripts
This is necessary because of how local changes to the schema interact
with the remote one.

When developing these locally, it is possible to import files from the
local filesystem instead of over the internet (i.e. add `./core' to the
`imports' attribute of a LinkML schema, instead of e.g.
`http://localhost/saved/schema/core'). This works fine for coding up the
schema but if these locally-developed YAML files are pushed to the local
web server, and referenced by child schemata, there is occasionally
strange behaviour with conflicting imports/prefixes.

The easiest option, then, is to just use the local webserver always, so
when building these, copy the latest YAML components to the local web
server.