Move as much as possible from copy scripts into makefile
Previously, I had the RDF generated only appear in the webroot when
running these scripts, since there was a neccessity to generate it
separately, due to dependency on these web resources when LinkML makes
the RDF conversion of the ontology, which it doesn't have when producing
most of the other formats (including Owl).

The idea now, which makes better sense, is that running `make(1)' will
produce a directory which can be copied to the web-root, with everything
we expect. This still has a separate, `site' rule which does everything
except generate/copy the RDF, then the default rule `all' runs the `site'
rule and deals with the RDF.

The installation / preparation scripts still exist, to copy that
directory to well-known places, or to prepare the target directory with
new YAML schema files (for local testing of changes).